Added send of mocapframes via udp
A mocapframe is serialized using googles
protobuf implementation. Inside the Model.Gen
folder a new MocapFrame class is generated
using the model.proto in the project folder.
To connect udp clients and the server new
config values are added. The server runs on
one port. All clients have to be set inside the config
as well (udp.receivers). A working example is
added inside the test class "TestUdpServer".